[fix] bug edit metier [refactor] on sort le fonction modal
